Implement actions framework, intended for use in merging events and analyze commands.
Implement SAMPLE_LANDSCAPE as a demonstration of actions framework.  Can be scheduled as an event, as well as called from Analyze mode (meaning it is now available in analyze, where it was not before).

Move cAnalyze instantiation into cWorld.  Lazily initialized on first use.   This is to support Actions using analysis state at any time.

Fix cEventManager to prevent crash when supplying a non-existent event name, error message should now be reported correctly.