Fix cTools::MkDir to properly close file handle after checking directory so that it no longer leaks file handles.
When printing dominant genotypes to the archive, MkDir gets called relatively frequently (under the new cDataFileManager scheme) and was chewing up the maximum open files.   At a frequency of 100 updates, this took over 100k updates and was not noticed.   With both Dominant and DominantParasite active, it happened around 50k updates and caused trouble.