Implement task accounting in cMutationalNeighborhood.   This essentially completes the implementation.
Also comment out some unused portions of cEnvironment and cReactionResult.  The side effect of the extra loop appears to be limited to modification of the receive_tasks bool array in cReactionResult.  As this array is not, and cannot, be accessed outside of cReactionResult at this time, it is fruitless to perform the tests.   The removal of instantiation, maintenance, and destruction of these unused arrays yields about a 9% performance improvement in single threaded landscape computation (given the default 9 task environment).  If this second loop becomes necessary in the future, a better method of determining which tasks to reevaluate should be devised.