Adjust cHardwareStatusPrinter/cHardwareTracer setup to be more generic, removing hardware specific method wrappers and dynamic casting.
Cleanup various hardware specific sections within cZoomScreen.  Added more base class methods attempting to remove the necessity for casting references to the base class.   There are still two places where cZoomScreen assumes that it is working with a cHardwareCPU object.   One is a fairly safe assumption (within the type specific UpdateCPU method).    The second is unsafe in that even though the option to view components is not advertised on the SMT/4Stack component zoom, it can still be accessed via the assigned key.   In debug mode this will cause an assert, the behavior in release mode is undefined.