Add in a true copy of SMT named cHardwareTransSMT.    Also reconfigured the support/config files to more clearly identify classic, smt, and transsmt files.  Updated the hardware classes to properly report their default instruction set file names.   Improved cHardwareManager so that it will load the default instruction set when INST_SET is '-', and also report that it is doing so as a comment.   Modified cAvidaConfig to use '-' as the default setting for INST_SET.   The net result is that to change cpu models now only requires two settings to be changed by default (HARDWARE_TYPE and START_ORGANISM).
Updated the consistency test settings so that they point to the new config file names.   Also added setting to the SMT test so that it sets the SIZE_MERIT_METHOD to what is expected by the tails, even though avida.cfg uses a different default.  We will have to decide which is the desired setting in the future, but for now 4 is the default for all standard configs (removed separate avida-smt.cfg).