Add cArgContainer and accompanying cArgSchema.  cArgContainer serves as a fast access container for arguments that have been parsed from an input file.  The parsing routine loads in named arguments with various types, the details of which are encoded in a cArgSchema object.   Schema entries specify the name of the argument, the target index, and a type specifier or default value.  When a default value is specified, the argument is regarded as optional, otherwise the entry is treated as a require value.
cArgContainer objects can only be constructed through the class static cArgContainer::Load method, which ensures that the supplied argument string is compliant with the supplied schema prior to returning.   Optionally, a tList<cString*> can be passed in to the Load method.  When a list is supplied, strings describing any errors that occur in attempting to load the arguments will be appended to it (this will simplify displaying errors in the future, where multiple worlds may be running and/or within a GUI).


Adjusted cTaskEntry (and references within cTaskLib) to hold a cArgContainer in place of the former info string.  New tasks can now use cArgContainers to parse the info string into fast access arguments (to speed up evaluation at each IO).

Changed matchstr task to use cArgContainer, as it was using the info string.