Added a new BIRTH_METHOD (7) that allows a parent to choose where it wants its offspring to be placed, based on that parent's facing.
Changed some defaults in avida.cfg:
 WORLD_X 60   (was 100)
 WORLD_Y 60   (was 100)
 VIEW_MODE 1  (to make the text viewer start in map mode; was 0, blank screen).
 BIRTH_METHOD 0 (Random in a structured environment; was 4 - mass action)

Minor code cleanups:
Fixed an unsigned int to int comparison in cMerit that was throwing a warning
during compilation.

Fixed an assignment-in-while-condition warning in cAnalyze.cc (added an extra
set of parentheses around the assignment, as the compiler requested).

Updated KNOWN_BUGS list with some more, useful cleanups that will need to wait
until we're willing to break backward compatibility of config files again.