Added a counter to each deme that keeps track of how many organisms are currently alive in that deme.
Created a new instruction called 'spawn_deme' that will choose a random
individual in the current deme and spawn it into another deme chosen at
random.  This process entrails wiping out all indvididuals in the target
deme so that the spawned indivdual has no compition there.  This is being
created as an instruction, but I expect it will usually be used as the
result of a task; when the deme completes the task, it gets to spawn.

Removed the DIE_PROB setting in avida.cfg.  This was supposed to give a 
probability of death whenever the die instruction was executed, but this
same effect could be done by making the die instruction have a chance of 
failure (which is working fo all instructions currently).

Plus some minor code and comment cleanups...

Up Next: I'll be cleaning up the task code and creating some deme-level
tasks to associate the spawn-deme instruction with.