Added in source/utils/process_map/ with a new tool to help generate movies of an Avida grid over time.  To compile, simply run 'make' in the directory (we should eventually tie this into our regular build system).  More directions forthcoming.
Fixed a bug in the Probabilistic scheduler where it crashed if all of the
organisms in the population die out.

Fixed a bug in the DumpGenotypeIDGrid where it was swapping the rows and
columns in the grid (this is hard to notice with square grids...)

Fixed a bug in the zoom mode of the text viewer where Avida would crash if
zoomed into an empty cell.

Fixed a bug in cView in the text viewer where Avida would sometime crash
when unpausing.