Merry X-mas!
1. THROW/CATCH

Corrected the behavior of the 'throw/catch' instructions
to my original intent. A 'throw' instruction jumps to
the next 'catch' instruction followed by a modifying
label that matches the 'throw' label OR ANY TRUNCATION OF
THAT LABEL, i.e. a catch without any modifying NOPs will
capture any 'throw' instruction, no matter what it's
label. 

(Note: I've never seen the label on 'throw' used by an
evolved organism, though it works correctly in tests. It
might be worth trying the opposite -- where a 'catch'
stops only 'throw's modified by truncations or substrings 
of the 'catch' label.)

2. GOTO/LABEL

Added a similar 'goto/label' suite of instructions (goto,
goto-if=0, goto-if!=0). A 'goto' instruction jumps to the
next 'label' instruction after it with an IDENTICAL 
modifying label.

(I'm starting runs to compare these alternate methods
of flow control.)

3. TRANSPOSON

Added a first-generation 'transposon' instruction that
adds insertions of itself with some probability to a
child genome when it has been executed in the parent. This
only works with 'repro' organisms.