==Test CPU Resources
Added an additional mode for resources that can be invoked from analyze mode. In place of a 0 or 1 for the [use_resources] flag in all commands, you can also use 2.

This implements updated and depletable resources, meaning that (1) updates pass in the test CPU -- it will reset resources according to the file whenever it finds a matching update and (2) these units can be used up by an organism. For test CPU purposes, we assume that an update is as long as the average time slice.

You can also specify an offset (in a number of CPU cycles) from the update that you specify for resources. This allows finer control over where organisms start a resource cycle. I added this as a parameter after the filename in the LOAD_RESOURCES command to avoid making the syntax for other commands even messier.

When invoked during a normal Avida run, the test CPU now supplies non-depletable resources in the initial amounts specified for the run. It had been showing no tasks executed whenever an organism tried to complete a task that required a resource in the archive/*.org files.

==Trace File

If you use instructions with an additional time cost (that ages the organism as if it had used more CPU cycles, but doesn't take any longer to execute), then the trace file also includes the actual CPU time that the organism has used in addition to it effective time that it has aged.

I also threw in a quick and ugly print out of current resources levels in the trace file when you use resource option 2 above.

==Successful reactions can clear the input buffer

I was becoming annoyed with test organisms saving an output value to "put" repeatedly in some pilot studies, so I added a REACTION flag (:clearsinput=1) so that successfully completing a reaction can clear the input buffer. It's implemented in a way such that if an IO or put completes one of these reactions then no further reactions that the output might also trigger are rewarded (rather than the standard behavior of rewarding all applicable reactions for an output).

==Misc Instructions

Added a new flavor of "sense" and some flavors of "put" and "IO" with "repro" built in at the request of my organisms.