Implement a full complement of wrapper objects for various threading constructs, include cMutex, cConditionVariable, and cRWLock.  All of these, as well as cThread itself now, have been written to support three different compile modes (pthread, Windows threads, and disabled).   The code for Windows threading has been written, but it is impossible to test it at the moment as some of the features require Windows Vista (earlier version do not support conds or rwlocks).   The default on windows is to still disable threading, but using the new wrapper objects rather than pthread construct overrides.
Add platform.h to tools.  This header file handles determining various platform features so that outside of it conditional compilation can be standardized on the macro function AVIDA_PLATFORM(FEATURE).   For example, to test if Avida is being compiled on windows #if AVIDA_PLATFORM(WINDOWS).

Fix cAnalyze::m_ctx so that it is a reference rather than a copy.  This should fix all of the LandscapeActions so that they work on batches in analyze mode, as expected.