Now for why some tests were reset.
Change behavior of cGenotypeControl::Remove to check if the genotype being removed is the current coalescence point. If so, clear out the coalescence pointer so that it will be reset on next call to UpdateCoalescent. 

Note, this is a workaround for the fact that our current implementation of coalescence assumes that there will only ever be one ancestor.   This really needs to be reworked, as there are numerous cases where such an assumption will be incorrect.