Added several undocumented detail command output options to the documentation.
Created a test for all detail output options (except those requiring sexual genotypes or forcing landscaping), in both .dat and .html formats.  It is long.

This test was created WITHOUT Charles' recent changes to analyze files in order to serve as a consistency test for those changes.  Therefore the expected results may need updating if those changes break consistency but in an acceptable way.