Setup cFlexVar to handle assignment and relationship operators (including a macro framework to semi-easily add more operators, but its hard to get it to be too simple with all of the combinations of base types needing to be accounted for).
Setup analyze.cfg to automatically be created if it does not exist on startup of analyze mode
(along with a warning message).

Added a "TEST" command to analyze mode.  This calls the method cAnalyze::PrintTestInfo(), which developers can fill
out with any test info they want (there is no guaranteed behavior other than it gets called when you run TEST in
analyze mode).