Added commands to convert the base type in cFlexVar (MakeInt(), MakeChar(), etc.)
Added a proper comparison command in order to compare two cFlexValues and produce a change quality.  This change
quality will them be use when we print a value to the screen to indicate if it is an improvement or if it makes
things worse.  This has NOT been hooked in anywhere yet, but I wanted to check it all in while everything was
still working.