Added a new FILTER command to analyze mode.  With this command, you can filter down the genotypes in the batch by any metric of your choice.  For example, if you do:
  FILTER fitness >= 1.5

This will limit the genotypes to those with a fitness of at least 1.5.  You can base this on any
value available in the DETAIL command.  So if you only want those organisms that can perform EQU,
you would do:

  FILTER task.8 > 0

...and there you have it.

I've also updated the documentation to reflect this new command.


Other stuff:
Trimmed back further on the use of tDataCommand templates, replacing them with the more general
cFlexVar where possible.

Fixed member variables in cFlexVar to begin with m_ to match the coding standard for Avida.

Changed analyze command "TRUNCATELINEAGE" to "TRUNCATE_LINEAGE".  Old version still works, but is
depricated.