Whitespace change in PrintActions.cc.
Created a utility to generate an up-to-date actions.html.  This utility reads
the ***Actions.cc files in the source/actions to find the names of all actions
that have been defined.  It then combines this with information from
source/utils/make_actions_html/actions_source_info to create a new version of
the actions.html file.