Fix cDataFileManager automatic output path creation to properly create the path as expected.  Also ignore relative path prefixes.
Enhance the test cpu and cCPUTestInfo to allow for changing the instruction set used in the test cpu without altering the rest of the world.

Fix cAnalyzeGenotype's usage of its internal instruction set to properly change the test cpu's inst set for all calculations.  This fixes a bug whereby map tasks crashed if not column arguments were specified.