Restructure global object registration, including the driver objects, the cActionLibrary instance, and the cAnalyzeGenotype data entry dictionary.
Add new tDMSingleton class that can be used to wrap and initialize singleton objects that self register with the driver manager for destruction on termination.

Add new tDataCommandManager, to wrap the functionality of data entry dictionaries (cAnalyzeGenotype is the only one at the moment), so that such structures can be managed by the new tDMSingleton class.