Add support for setting the state grid to use in the TRACE command.  This can be accomplished by passing "sg=#" as the first argument following the output directory.  For example:
TRACE archive/ sg=1

As in the example, # is simply an integer at the moment, where 0 is the first grid listed in the environment file.   Eventually this should be changed to use the name of the state grid.

Also note that since the TRACE command creates the file name based on the organism's ID, it will overwrite existing files in the output directory.   When performing analysis on a genotype with multiple different state grids it will be necessary to specify different output directories.

All other arguments to the TRACE command should be processed in the same order as before, with and without a 'sg=' argument.