Implement the capability to toggle the expiration activity of the I/O -expire instructions.
Add a new action SetConfig that will allow a config option to be set in the events files.

Cleanup and fill out the mutation types in the configuration file and support actions.  Note that newly added mutation types have not been implemented yet.  The config options for such mutation types are currently ignored.   This will be corrected in a future commit.