Fixed documentation for SetMutProb and ModMutProb.
'Cause, see, I wanted my events.cfg file to be readable, so I typed in the defaults instead of letting them default.  But the default mut_type value in the documentation was actually not a valid option at all, and there went 13 sets of 50 runs each, down the drain.

So I've fixed that, and added what the other options for mut_type are.