Add support for toggling during-run parent distance tracking. When TRACK_MAIN_LINEAGE is enabled these calculations are unnecessary since recalculate in analyze mode can and already does perform these calculations again.   The avida.cfg option TRACK_PARENT_DIST will default to disabled, however it can be enabled for experiments that want parent distance information but have disabled TRACK_MAIN_LINEAGE for performance reasons.
NOTE: This will break a number of consistency tests, since the 'parent_dist' column will be filled with -1 by default.   Although it would be preferable to actually remove the column altogether, certain code currently depends on the exact ordering and number of columns in the default dump files in order to work correctly.   Future versions of Avida will fix this.