Adds an iterative mergesort to tArray, along with a simple unit test.
Reasoning: it seems reasonable to wish to sort a tArray.  The only current way is to use tArray::QSort, which wraps qsort.  However, since qsort may have different implementations, it may not always sort elements of the same value in the same order, therefore possibly breaking consistency in the case of, oh, sorting phenotypes by # of cpus -- two different phenotypes with the same number of cpus might appear in different orders.  Therefore it is necessary to have our own sort.
The sort I am most comfortable with is mergesort, and I'm definitely not up to writing a well-optimized quicksort.  Therefore I wrote a mergesort, which I did not optimize but should serve well enough.

QSort ought to be perfectly fine to use when sorting tArrays not composed of elements which have this particular problem (same value for sorting, but actually different).