Parasite genotypes are now output by SaveStructuredPopulationBG.
Enhance cDataFile to support setting filetype and tracking column format.  This allows for data file output hand off to sub objects.   Not quite ready for structured data save, but it is a step in that direction.  For example, cBGGenotype handles outputting its relevant information and defers outputting cMetaGenome information to the cMetaGenome class now.