Changed the translation of a nop-specification into a resource to be collected.  Avida will no longer crash if, when using collect-type instructions, the number of resources is not a power of the number of nops.  This does not apply to the TransSMT version of the collect instructions.
This breaks consistency for older collect tests (although not qualitatively).  Expected results deleted in preparation for new result generation.  Non-collect-associated consistency was not broken.

New tests added to check the workingness of the new specification for various numbers of resources.

Release notes updated with this change.

(The cCodeLabel change is just some extra comments.)