blwalker@msu.edu: Fixes the issue where resource id in ResourceLib does not necessarily correspond with resource id in ResourceCount.  Adds MergeResourceAcrossDemes action.
MergeResourceAcrossDemes sums the named deme resource across demes, sets the named global resource to the sum level, and zeroes out the deme resource.  Inflow and outflow are also merged -- inflow is summed and the outflow remains the same.  The global resource must already exist in the environment file.

To support lack-of-code duplication across MergeResourceAcrossDemes and several other resource-related environment actions (set, setinflow, setouflow, etc.), functionality was moved up to cPopulation.

While implementing this action, it became clear that a resource's (environmental, ResourceLib) ID was a bad choice to use for indexing into a ResourceCount, since a ResourceCount may not contain all resources (e.g. if some are deme resources and some are not.  This problem was previously solved (twice, redundantly) by implementing looping functions in cResourceCount which associated a named resource with its ResourceCount id -- or index.

There is now a non-looping solution.  When the environment is initialized, a resource is given an index as well as an id.  The id remains as it always was.  The index refers to its place in its ResourceCount -- essentially the deme resources and non-deme resources are assigned indices separately, since they will exist in separate resource counts.

The two redunant methods of cResourceCount, GetResourceCountID() and GetResourceByName(), will be removed in the future, as the looping solution is inferior to a simple index lookup.