Implements copy constructor and operator= for cGenotypeBatch.
Additionally, adds deletion of pointer members to the destructor.

The implementation of operator=() is broken in this commit and
will usually cause a segfault.  It is fixed two commits down
the road.

It's been a long time since I've written copy constructors;
someone else might want to take a look at this commit and the
one that fixes it to make sure I haven't done anything terrible.