Minor fixes throughout deme-related code
Adjusted feedback mechanism for many actions to function properly using
the passed in feedback object as opposed to the uninitialized world
object.

Setup cHardwareCPU::DoSenseFacing to not assume that there will always
be a cell in front of the executing organism.

Commented out code associated with resistance from steep hills; the
code was assuming that there was always at least one resource and
crashing when this was not the case.

Set default duration for events (which setup targets) to be 0, not -1
to prevent crashes.

Lots of tabbing and formatting changes.