Fix Avida2Driver Feedback & Prepare for AvidaEd WebUI
Some of the changes to Avida's UserFeedback object to make it
queue responses (for asynchronous purposes) weren't passed on
to the Avida2Driver's UserFeedback struct.  There may be
more drivers that need updating, but so far I've only tested
the "primitive" driver.

I modified the build_avida script to take a new parameter: 'avida-ed'.
'avida-ed' will trigger the Emscripten build of the avida-ed web
UI driver and associated resources.  The build system is still
a mess.  Associated CMakeFile.txts have been updated as well to
account for this new functionality.

A bare-bones driver is now in place for communicating via JSON
with requesting JS entities.  The actual code for this is not
in place (not even stubs), but the driver initializes, and exits
main.  This is what we currently want since we're transferring
start control to the calling JS entity.  Emscripten will allow us
to terminate main and keep the avida-core active, awaiting additional
commands.