I'm replacing rapidjson with another json library
This new library is simply called json.  It is at 1.0-rc1 currently,
which should be sufficient for our purposes at the moment.

The *major* benefit of this is that it makes dealing with JSON data
incredibly easy, it uses modern c++ features (so we do need c++11 or
c++14), and is overall a better fit than the micromanaged rapidjson.