Remove 16-bit width for m_cycle_count and m_last_output fields for GP8, Experimental, and BCR cpu types.
Avatar regression tests were exceeding the 16-bit limited placed
on the m_cycle_count field when executed against a release build.
An assert that limited cycle_count to a width of 16-bits was
triggered with a Debug build, but has not been detected for
some time.

As these bit-widths seem to be memory optimizing in nature, I'm
removing the bit-width limitation from these virtual CPU models.

This will also fix the silent corruption that the avatar environment
has been making by exceeding the bitwidth boundary with assertations
disabled.