Make avida.js path substition POSIX-compliant.
When AvidaED-web builds, the path for locating the preset members
of the virtual file system (from apps/viewer-webed/resources) has
an incorrect path hard-coded into avida.js.  Sed is used at the
end of the build process to modify avida.js to contain the proper
paths.  However, inplace substitution using the -i flag is not
consistent between different OSes and is not part of the POSIX
spec for sed.  The current modification relies on creating a temp
file with the proper changes and moving it to the proper location.