Added PrintSpatialResources action.
The current print event system for spatial resources was lacking
an easy way to obtain the current resource abundances per cell
for spatial resources.  An existing function proports to allow
an "easy-to-import" output designed for matlab.  However, the
file format is not something that can be imported using python
related matlab-loading utilities.

The new PrintSpatialResources action will print a space-delimited
file containing the update, the resource name, and the abundance
count for all cells.  All resources are contained within this
file and will require post-processing to separate them out.

The file may be read in using Pandas for Python.