Refactored Resource Upating; fixed cDeme bug
The main function call to update resources was split to more clearly
differentiate how spatial and non-spatial resources are updated.

Documentation about the update process was added.

I added a helper method in cResourceCount to tell if a resource is
spatial or not and applied that method wherever I could.

cDeme had a delete[] bug.

All tests passing.