Add Feedback to all Actions
All action classes currently have the ability to receive a Feedback
object, which is used to pass information to the user during the course
of the experiment.  However, use of this parameter was not widely used
even in cases where feedback would be useful -- such as when an action
fails.

cAction now stores a reference to a Feedback object.

Cleaned up white space a little in the process.

All tests passed.