Fix RESOURCE cellbox math; unmask actual cell vals
I rewrote the edge handling of the RESOURCE cell boxes after finding
issues with the last implementation.  This one appears to work --
especially after removing the "masking" that was done in the
cResourceCount API to retrieve cell resources.  Prior to this commit, I
had made anything not in a cell box return empty for all resources
(global and spatial), but in this commit I re-enabled getting the actual
value of the cell abundance rather than returning empty to make sure
there were no weird leaks.